Removed wait in installer when finshed
When the installer had finished installing the files it would wait on the log file page.
It is not obvious that it had finished installing and that the user had the press the next button.
Removing this line makes the installer jump to the finished page automatically when it has finished installing making it very obvious that it has finished.