Added a stylable svg widget and button
This adds a custom SVG widget that can be styled via the widget background.
This allows changing the style via style sheets rather than code making style sheets more useful.
It also adds a button containing a stylable SVG widget and the same with a tool tip.
The latter should be refactored to work with any class and remove the need for a subclass.