Refactored plugin dependency
Plugin could originally only define other plugins as dependents.
This refactors the dependency system to include python dependency and the libraries distributed with python, other python libraries and plugins.
This will allow us to update libraries and know if the plugin is compatible.
Updated the import validation to verify if the imported module is listed in the dependents.
Still need to validate this when the plugin is enabled.