Added an argument to trace function calls
PySide6 can be a bit unstable. If the wrong object is passed it will usually crash with an access violation rather than raising a python exception which makes finding the origin rather hard.
Adding the --trace command line argument will print the qualified name of each function as it is called. If the program hard crashes the last line should be the origin of the crash.