Fixed access violations
When the render widget is closed/destroyed the render loop needs to be stopped by setting the root entity to a null pointer.
The __del__ method is not called until after the C++ instance has been destroyed which was causing the access violations.
I tried getting other destruction methods to work but I could not get any to work.
The best solution I found was to set it to null each time the widget is hidden (which happens before it is destroyed)
It should also destroy the aspect engine but I can't work out how to run code before the widget gets deleted.