Enable and disable the render aspect
This registers the render aspect on show and disables it on hide.
If the render aspect is still registered when the C++ instance is deleted a warning is displayed in the console.
This would usually be done when the instance is deleted but I can't work out how to get code to run before the C++ instance is deleted so this will have to do.