Removed the Qt3D code
I experimented with Qt3D but it was not very performant with the number of entities that we would be throwing at it.
There were also virtually no examples and the documentation was lacking making it very difficult to develop for.
For these reasons I think we will stick with direct OpenGL.