Refactored plugin hooks
Plugin hooks were previously global in the root module of the plugin.
If we wanted to change the hook format it would have been quite difficult.
This changes the plugin format so that there is one root attribute called `plugin` which is an instance of PluginV1.
If we want to change the hook format in the future we can implement another class while still supporting the old format.