Added a thread_manager plugin
If a thread is not stopped before the application exits it will crash the program.
Calling quit and then wait for each thread is the general suggestion but this means the program needs to wait for each thread to finish before quitting the next thread.
This plugin automatically quits all threads and then waits for all threads which means that all the threads can finish in parallel.