Fixed some OpenGL issues with the window system
The OpenGL destruction was not being run which caused Qt to do a very slow teardown.
OpenGL destruction is now run when the widget is hidden because this is the only way I have found to do it when the widget may be destroyed before the containing window.
This also means that we must do OpenGL initialisation in the showEvent to set it back up after it gets hidden.
Fixed a number of crashes in a number of different situations.