Added a QOpenGLContext C++ extension
The QOpenGLContext implementation does not allow access to the context after the destructor has started.
This means that the context cannot be made current from within aboutToBeDestroyed rendering it worthless.
This adds a Cython wrapper for the C++ context instance so that it can be made current after destruction has started.