Added 3D world viewer (#86)
* Fixed get_game_version call

* Open the level after getting it if not already open.

* Fixed typo

* Expose View3D to other plugins

* Added chunk builder stub file

* Added View3D layout and button

* Expanded Optional

* Added error chunk and null chunk geometry

* Renamed WidgetLevelGeometry to LevelGeometry

* Renamed GlData to CanvasGlData

* Added debug logging and comments

* Renamed WidgetChunkData to ChunkData

* Fixed LibraryUID hashing and equality.

* Split up default layout plugin

* Renamed package

* Implemented missing widget population

If a layout is added before a required widget is added it will be displayed as a missing widget.
Once the widget is added it can be populated.
Likewise when plugins are disabled.

* Added main window destruction

When the plugin is disabled, the window will be destroyed.

* Reformatted

* Removed Drawable ABC

* Renamed modules

* Added prototype renderer

* Removed multisample

* Updated stub file

* Reformatted

* Fixed issues starting and stopping threads

The stop method is called twice without being started for some reason.

* Drop chunk meshing to idle priority

This should make it lag a little less.
Releasing the GIL will also improve this.

* Reduced thread count

I may increase this again but they are all locked by the GIL currently.

* Remove duplicate code

* Renamed OpenGLResourcePackHandle

* Remove RPC

* Moved the texture out of the chunk

If each chunk stores the texture, we need to re-bind the texture for each draw making it slower.
When the resource pack changes we just unload all chunks like Minecraft does.

* Set texture location once

* Refactored chunk mesh module

* Reformatted

* Update references

* Added first chunk mesher

* Refactored and improved chunk meshing

Moved performant code to C++ so it does not need the GIL.
__init__ is now extension

* Moved C++ chunk meshing code

Moved pure C++ meshing code to its own file.

* Removed Amulet:: prefix

* Refactored chunk mesher

Added neighbour chunk component inputs. This is needed to determine culling at chunk edges.
Merged the chunk mesher into one function.
Converted storage to raw pointers.

* Modified python function signature

* Refactored chunk mesher

Switched to a transparency array. This makes the code a bit simpler and reduces the number of calls to get_block_mesh which should make it a bit faster.

* Fixed mesher call

* Fixed stride bug

* Fixed transparency issues

Draw the blocks with translucency after the opaque blocks.

* Implemented vertical culling

* Pencilled in neighbour chunk getting

This causes a deadlock currently.

* Implemented getting neighbour chunks

* Implemented neighbour chunk culling

* Implemented block shading

* Updated widget titles

* Reformatted

* Bumped amulet dependency

* Moved tests

* Added tests init py

* Added a dummy test

* Added missing import