Add a function to modify sys.modules
The modules mapping is stored in sys.modules and in the C API and used from both places.
To truly override it we need to set it in both places.
There is no public API to set it in the C API so some janky code is needed to dig around in private memory.