Refactor amulet.app (#95)
* Rename amulet_editor to amulet.app

The amulet_editor module is no longer the actual editor so this name is confusing.
The editor is implemented through plugins.

* Move resource module and rename get_resource_path

Move and rename amulet_editor.resources to amulet.app.resource
Rename get_resource to get_resource_path to make it clear what the function does.

* Move the paths module

* Move plugin modules

* Move locale modules

Move locale modules into amulet.app and made objects public.
Remove amulet_team_locale plugin.

* Rename Translator class

* Rename some amulet_editor references to amulet.app

* Swap custom path handling for get_resource_path

* Shorten CLI name

* Move CLI modules

* Move the exception module

* Improve exception handler

* Rename module

* Move promise module

* Move and refactor signal module

* Remove the RPC module

* Move the trace module

* Remove empty modules

* Remove unused _app module

* Move the invoke module

* Fix imports

* Move the application modules

* Skip plugin unload if the module failed to load

* Add dependency

* Rename the editor command

* Move DragContainer class

* Move icon and label modules

* Add missing requirements

* Reformat

* Add a function to modify sys.modules

The modules mapping is stored in sys.modules and in the C API and used from both places.
To truly override it we need to set it in both places.
There is no public API to set it in the C API so some janky code is needed to dig around in private memory.

* Improve plugin management

Set sys.modules in both locations.
Refactor the code a bit

* Add missing dependencies

* Move the appearance module

This should get re-implemented at some point

* Update stub generator

* Reformat

* Fix dependency resolution