Remove promise and refactor the renderer (#98)
* Switch to progress manager functions

* Simplify resource pack plugin

The API for this plugin was far too complex.
The new implementation is far simpler and fixes a number of race conditions.

* Remove promise from create_atlas

* Remove promises from gl resource pack

* Add return type hint

* Refactor resource pack plugin

* Refactor opengl resource pack module

* Remove unused import

* Refactor level geometry

* Remove promise module

this has been replaced by the task manager classes

* Improve debug message

* Add debug logging

* Add debug logging

* Ensure functions are called from the main thread

* Fix context destruction

The bound method was not being called for some reason I don't understand but a normal function is called just fine

* Remove dynamic_cast function

This is implemented in the utility library

* Fix invalid QOpenGLContext

PySide6 breaks the link between the Python and C++ objects before they are fully destroyed.
For the QOpenGLContext this is before aboutToBeDestroyed is called.
The C++ object is still valid so we can construct a new instance with the stored pointer and use it until it is actually destroyed.

* Store OpenGLResourcePackHandle weakly

* Add debug message

* Remove old import and comment

* Add missing imports

* Reformat

* Add debug message

* Refactor level geometry

* Rename _gl_data attributes

* Move event binding into show and hide

* Split colour variables

* Add debug logging

* Manually garbage collect when unloading plugins

For some reason some objects are not automatically garbage collected.
I have tried resolving this without manually running the garbage collector but this seems to be the only way.

* Improve resource pack destruction

* Improve comment

* Improve level geometry

Improve threading.
Improve cancelling.
Improve OpenGL initialisation. Chunks are initialised in batches rather than individually.

* Replace processing_chunks to bool

There was no reason for this to be a set when it can be a bool in the ChunkData class

* Set default cache dir

* Remove unused variable

* Fix thread locking issue

The manager thread could be woken and go back to sleep before the worker has exited