Improve widget framework (#117)
* Use human readable identifiers

UUID identifiers are not particularly human readable

* Improve documentation and errors

* Require explicit initialisation of main window

This stops plugins interacting with the main window API if it is not being used

* Fix main window destruction

* Restructure the window package

* Fix imports

* Remove abstract tab engine classes

* Fix import errors

* Rename view_container to splitter

* Reformat

* Simplify RecursiveSplitter

* Fix mypy error

checkedButton can return None. This was not documented in the stubs.

* Add prototype widget base class

* Add improved tab widget classes

* Add WIP tab dragging

* Reformat

* Fix mypy warnings

* Remove unused features

* Split and rename tab widget getter

* Fix tab enabling when removing the active tab

This finds another tab to activate before removing the tab

* Fix window border error

If the cursor is in the shadow of a widget it will be considered inside the widget but it is not actually in the geometry.
Handle these cases.

* Move the tab by one pixel

The tab could interact with cursor blocking interaction with other widgets

* Add tab dropping

* Mark tab getter as final

* Remove create method

* Remove unused import

* Move private data into a wrapper class

* Improve and integrate new tab system

* Remove fusion styling

I don't like how checkable buttons look in the fusion style

* Finalise layout system

* Reformat