Improve exception handling in the canvas
If an exception occurs in initializeGL, the other opengl functions will not run.
If an exception occurs in paintGL only the first occurrence of each unique traceback will be displayed in a dialog. All exceptions will be logged. This is to stop an excessive stack of exception widgets building up.