Improve error handling (#118)
* Add an error handler

If an exception is raised in a function bound to a signal, qt will ignore the exception.
This adds an error handler to display those errors to the user in a dialog and adds them to the log file.

* Improve the exception dialog display functions

display_exception does not block the event loop. It is shown and removed when the user clears it.
Make sure display_exception_blocking is called from the main thread.

* Improve exception handling in the canvas

If an exception occurs in initializeGL, the other opengl functions will not run.
If an exception occurs in paintGL only the first occurrence of each unique traceback will be displayed in a dialog. All exceptions will be logged. This is to stop an excessive stack of exception widgets building up.