Add support to cancel loading resource packs
Fixes #101
When the canvas is being destroyed if the resource pack is still being loaded it will cancel the loading and wait for it to finish.
This also adds the ability to cancel loading a resource pack if another one is being loaded so it doesn't have to wait for the first to finish only to discard it.