Improve toolbar (#165)
* Add toolbar position customisation

The toolbar position can be changed with a variable.
Perhaps in the future this can be configurable in the GUI.

* Add global variables

This makes it easier to understand what the numbers mean.
Perhaps in the future this can be configurable.

* Remove unused lock

* Swap to QWidget and simplify constructor

* Make DragContainer scrollable and improve dragging

The dynamic button container is now scrollable.
Improve dragging behaviour. The previous implementation was a bit fragile and broke a little with the switch to the scrollable widget.

* Rename classes and variables

* Remove arg doc

* Add arrow buttons and fix highlighting

Add arrow buttons when the widget gets too small for those that do not have mouse wheels.
Fix button highlighting when scrolling. When scrolling the buttons do not get the mouse leave event so do not get unhighlighted.

* Fix dragging and scrolling

If scrolling while dragging, highlighting should not change and the dragged widget should move to the new location.