Make all toolbar labels visible while hovering on the toolbar
Previously only the currently hovered button would show the label.
I think it is better if all labels are visible while hovering on the toolbar.
This also works when scrolling and/or dragging.
Switch to a button for the label display because I prefer the style to a plain label and I couldn't find anything else better.